ajeitando diretorios e readme atualizado (acabei mandando o commit sem dar git add)
