+++ Saltitantes/src/Controller/SimulationEngine.java
@@ -1,10 +1,10 @@
 package Controller;
 
-import model.domain.Cluster;
-import model.domain.Duende;
-import model.domain.GuardiaoDoHorizonte;
-import model.domain.datastructure.TreeMapAdaptado;
-import model.domain.interfaces.EntityOnHorizon;
+import model.datastructure.TreeMapAdaptado;
+import model.entities.Cluster;
+import model.entities.Duende;
+import model.entities.GuardiaoDoHorizonte;
+import model.entities.interfaces.EntityOnHorizon;
 import java.util.ArrayList;
 import java.util.List;
