+++ Saltitantes/src/model/entities/Duende.java
@@ -1,6 +1,6 @@
-package model.domain;
+package model.entities;
 
-import model.domain.interfaces.EntityOnHorizon;
+import model.entities.interfaces.EntityOnHorizon;
 import java.util.Random;
 
 public class Duende implements EntityOnHorizon {
