+++ Saltitantes/src/model/entities/resources/sprites/duende.png
