+++ Saltitantes/src/model/entities/resources/sprites/guardiao-com-pa.png
