+++ Saltitantes/src/view/SimulationView.java
@@ -9,10 +9,10 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import javax.imageio.ImageIO;
-import model.domain.Cluster;
-import model.domain.Duende;
-import model.domain.GuardiaoDoHorizonte;
-import model.domain.interfaces.EntityOnHorizon;
+import model.entities.Cluster;
+import model.entities.Duende;
+import model.entities.GuardiaoDoHorizonte;
+import model.entities.interfaces.EntityOnHorizon;
 
 public class SimulationView extends JPanel {
     private static final int WIDTH = 800;
