+++ pom.xml
@@ -70,6 +70,7 @@
 
     <build>
         <sourceDirectory>Saltitantes/src</sourceDirectory>
+        <testSourceDirectory>Saltitantes/test/</testSourceDirectory>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
